+++ src/file_ext/async_impl/async_std_impl.rs
@@ -8,6 +8,5 @@ async_file_ext!(File, "async_std::fs::File");
 
 test_mod! {
   async_std::test,
-  use crate::async_std::AsyncFileExt;
   use async_std::fs;
 }