+++ src/file_ext/async_impl/fs_err3_tokio_impl.rs
@@ -8,6 +8,5 @@ async_file_ext!(File, "fs_err::tokio::File");
 
 test_mod! {
   tokio::test,
-  use crate::fs_err3_tokio::AsyncFileExt;
   use fs_err3::tokio as fs;
 }